requirements.txt
bidict==0.21.2
certifi==2020.12.5
click==7.1.2
dlib==19.16.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SocketIO==4.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
opencv-python==3.4.5.20
python-engineio==3.14.2
python-socketio==4.6.1
six==1.11.0
Werkzeug==0.14.1
wincertstore==0.2
